Only check for semi-sync ack request if semi-sync is enabled for this connection
I think this is correct, since the docs[1] for enabling semi-sync
indicate that not all slaves need to support it, even if the msater
does.  This does resolve the problem I sort-of-reported in #13, which
appears to be two bytes missing from the beginning of the packet at
the time the event is parsed.

[1]: https://dev.mysql.com/doc/refman/5.5/en/replication-semisync-installation.html