Added test showing that the decimal decode error.
I am getting some very very small numbers instead of the expected
numbers. The issue turned out to be the with the inital byte. When
there was a negative number that byte and the number was small enough
to fit in one byte, the initial byte size would be 1. We were then using
a 4 byte mask to mask out 1 byte value leaving the mask as part of the
value, causing the error. This was only an issue when the number was
negative as the mask was the 2's complement. For positive number the
mask is a zero, so it was not contributing to the final value.