Added tests for decode decimal and fix the implementation.
The implementation of the decode decimal was not maksing out the
compressed fields. Added a set of test based on the one from mysql.

This, also, fixes the decode decimal implementation to make sure they
pass these test. Include, is a binlog file containing the insert
statment, and a simple go program to generate the sql file.