Fix unclosed conn, shutdown errors
Closing a canal by calling Close() results in errors as described in #273 [1].
Fix that by explicitly ignoring context cancellation as it's not really
an error, but an expected part of Close().

Also fix the unclosed connection described in #361 [2] by using a
separate connection to kill the binlog streamer.

[1] https://github.com/siddontang/go-mysql/issues/273
[2] https://github.com/siddontang/go-mysql/issues/361

This closes #273.
This closes #361.