Initial work on supporting TLS in the driver
It's supported elsewhere in the codebase, but currently the driver has
no way to flag/indicate TLS through the DSN.

This uses additional parameters / query strings to indicate ssl. I'm
not sure _at all_ at the moment how to do the more complex custom
tlsConfig, but as an inital commit we can try doing basic TLS.

I could be being naive, but to me using `url.Parse` rather than this
custom parsing code would be easier. But perhaps there a reason why
it's done this way and why `?` is used instead of `/` for the database.
I've done my changes to fit in with the current way of doing things.