Reference commit - Custom TLS config
This was just so I could test something locally. I hardcoded the CA 
cert and domain I needed in and then had a custom TLS config which used 
the CA cert and domain only.

The `make([]byte, 0)` are effectively Nils and a way to set the 
cert/key pair to Nil.

Ideally what I'd like to do is finish this off properly, but I don't 
actually need this change. The basic TLS changes are all I need.