Allow multiple custom TLS configs in driver
Stored in a map using the address (hostname:port) as the key.

Initial work at addressing review feedback. Possibly/probably there is
mutex and locking stuff we want to do. Maybe overkill though?

Using the address vs the full DSN makes sense to me as the key. It's a
little cleaner and I can't see a situation where a slight difference in
the DSN (i.e. different database) would require a different certificate
or TLS config.

This is actually pretty neat as we just use `&ssl=custom` and it looks
up the TLS Config (that people will have to pre-register/store) based
on the address in the DSN.