Switch DSN style for driver, but still allow legacy form
Address feedback from review.

The old style of DSN uses a `?` as path separator. This made parsing 
the DSN more complicated than necessary especially when adding in 
parsing of url parameters.

By switching to the more standard `/` as a path separator we can use 
`url.Parse` to more easily split up the DSN and pull out the query 
strings.

However... for compatibility we should still allow the old style.

So what this commit does is:

1. Use a regex to identify whether a legacy or standard DSN style. We 
are assuming that if a "/" occurs after an "@" and no more "@" or "/" 
occur after that then it's the new/standard DSN style
2. Prefix the DSN with a scheme so we can use `url.Parse`
3. Only supports the url parameters (`ssl=X`) for the newer style DSN

I.e. this shouldn't break anything for existing use cases, but does 
also allow people to start using TLS in the driver.

There are probably edge cases in parsing of DSNs, but there would have 
been in the old way of doing it as well.