Split out DSN parsing into function and add unit test
Review feedback.

Needed to split out to add a unit test and that naturally resulted in
adding a struct/type for the connection info.

Ideally I'd go back in time and split out the original code and use
this unit test for that - and maybe that'll happen if deemed necessary
with some rebasing.

But if not we should be able to satisfy ourselves that the changes to
the driver parsing code are ok with this test.