Add basic column count mismatch to trigger table schema reload.
This was originally written by Doug Weber on 2017-01-03.
See https://github.com/MasteryConnect/go-mysql/commit/74b682cc618aae4351111ef757e0e38d1fa9e327

We are re-applying it now because of the significant amount of drift
between the upstream and the original fork. Since it was more than 5
years behind.