Add `VisibilityBitmap` to `TableMapEvent` for MySQL 8.0.23+ Invisible Columns
Changes:
- `TableMapEvent.VisibilityBitmap`
`VisibilityBitmap` is a bitmap where each bit represents the visibility of a corresponding column in a table.
If a bit is set, it indicates that the corresponding column is NOT an invinsible column.
Invisible column was introduced in MySQL version 8.0.23.

- `TableMapEvent.VisibilityMap`
`VisibilityMap` lists boolean values of which is true if column of the same index is NOT an invisible column.

Co-authored-by: sean <sean.k1@kakaoent.com>