replication: Make ServerVersion a proper string
The ServerVersion is a slice of 50 bytes. If the ServerVersion was
"9.1.0" then the rest of they bytes would be 0x0.

This caused the printout of the FormatDescriptionEvent in go-mysqlbinlog
to look weird.

Before:

```
=== FormatDescriptionEvent ===
Date: 2024-10-24 14:42:04
Log position: 127
Event size: 123
Version: 4
Server version: 9.1.0^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@
Checksum algorithm: 1
```

After:

```
=== FormatDescriptionEvent ===
Date: 2024-10-24 14:42:04
Log position: 127
Event size: 123
Version: 4
Server version: 9.1.0
Checksum algorithm: 1
```